requirements.txt
torch>=2.3.0
transformers>=4.45.0
mamba-ssm>=2.2.2
causal-conv1d>=1.4.0
flashinfer-python>=0.1.6
pynvml>=11.5.0
pandas>=2.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
pyyaml>=6.0
tqdm>=4.66.0
pytest>=8.0.0
